Refactoring and lint fixes
